+++ Blammo-wai/CHANGELOG.md
@@ -1,4 +1,8 @@
-## [_Unreleased_](https://github.com/freckle/blammo/compare/Blammo-wai-v0.0.0.1..main)
+## [_Unreleased_](https://github.com/freckle/blammo/compare/Blammo-wai-v0.0.0.2..main)
+
+## [v0.0.0.2](https://github.com/freckle/blammo/compare/Blammo-wai-v0.0.0.1...Blammo-wai-v0.0.0.2)
+
+- Support `Blammo-2.1.0.0`
 
 ## [v0.0.0.1](https://github.com/freckle/blammo/compare/Blammo-wai-v0.0.0.0...Blammo-wai-v0.0.0.1)