+++ Blammo/CHANGELOG.md
@@ -1,4 +1,26 @@
-## [_Unreleased_](https://github.com/freckle/blammo/compare/Blammo-v2.0.0.0...main)
+## [_Unreleased_](https://github.com/freckle/blammo/compare/Blammo-v2.2.0.0...main)
+
+## [v2.1.0.0](https://github.com/freckle/blammo/compare/v2.0.0.0...Blammo-v2.1.0.0)
+
+Removes less frequently used definitions from the main `Blammo.Logging` module
+into other modules.
+
+- Moved from `Blammo.Logging` to new module `Blammo.Logging.ThreadContext`:
+  `MonadMask`, `withThreadContext`, `myThreadContext`, `Pair`.
+- Removed from `Blammo.Logging` (still available in `Blammo.Logging.LogSettings`):
+  `LogSettings`, `LogDestination (..)`, `LogFormat (..)`, `defaultLogSettings`,
+  `LogColor (..)`, `setLogSettingsLevels`, `setLogSettingsDestination`,
+  `setLogSettingsFormat`, `setLogSettingsColor`, `setLogSettingsBreakpoint`,
+  `setLogSettingsConcurrency`.
+- Moved from `Blammo.Logging` to new module `Blammo.Logging.Setup`:
+  `HasLogger (..)`, `withLogger`, `newLogger`, `runLoggerLoggingT`, `LoggingT`,
+  `WithLogger (..)`, `runWithLogger`
+
+`Blammo.Logging.Simple` has been expanded to include reëxports of:
+
+- `Blammo.Logging.LogSettings`
+- `Blammo.Logging.Setup`
+- `Blammo.Logging.ThreadContext`
 
 ## [v2.0.0.0](https://github.com/freckle/blammo/compare/v1.2.1.0...Blammo-v2.0.0.0)